move link to fuser footer button
