+++ apps/web/src/components/sidebar/index.tsx
@@ -4,7 +4,6 @@ import {
   useAgents,
   useDeleteThread,
   useIntegrations,
-  useInvites,
   useMarketplaceIntegrations,
   useThreads,
   useUpdateThreadTitle,
@@ -477,42 +476,6 @@ function PrefetchAgents() {
   return null;
 }
 
-function InvitesLink() {
-  const { data: invites = [] } = useInvites();
-  const href = "/invites";
-  const match = useMatch(href);
-
-  // If no invites, don't show the link
-  if (!invites.length) {
-    return null;
-  }
-
-  const handleClick = () => {
-    trackEvent("sidebar_navigation_click", {
-      item: "Invites",
-    });
-  };
-
-  return (
-    <SidebarMenuItem>
-      <SidebarMenuButton
-        asChild
-        isActive={!!match}
-        tooltip="Invites"
-        className="h-9 relative"
-      >
-        <Link to={href} onClick={handleClick}>
-          <Icon name="mail" filled={!!match} />
-          <span className="truncate">Invites</span>
-          <span className="absolute right-2 top-1/2 -mt-2 flex h-4 w-4 items-center justify-center rounded-full bg-red-500 text-[10px] font-medium text-white">
-            {invites.length}
-          </span>
-        </Link>
-      </SidebarMenuButton>
-    </SidebarMenuItem>
-  );
-}
-
 export function AppSidebar() {
   const { state, toggleSidebar, isMobile } = useSidebar();
   const isCollapsed = state === "collapsed";
@@ -581,9 +544,6 @@ export function AppSidebar() {
                       </SidebarMenuItem>
                     );
                   })}
-                  <Suspense fallback={null}>
-                    <InvitesLink />
-                  </Suspense>
                 </SidebarMenu>
               </SidebarGroupContent>
             </SidebarGroup>
